% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csv2midi.R
\name{csv2midi}
\alias{csv2midi}
\title{convert csv file to midi file}
\usage{
csv2midi(csv_file, midi_file)
}
\arguments{
\item{csv_file}{path of a midi file used as input (character)}

\item{midi_file}{path of a csv file used as output (character)}
}
\description{
\code{csv2midi} takes a midi file and outputs a textual representation
in a csv file
}
\examples{
\dontrun{
csvfile <- system.file("mididata/Nokia_Tune.csv",package="midicsvR")
midifile <- "midi_from_csv.mid"
csv2midi(csvfile,midifile)
}
}
