% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_conversion.R
\name{file_conversion}
\alias{file_conversion}
\alias{midi2csv}
\alias{csv2midi}
\title{conversion between midi and csf files}
\usage{
midi2csv(midi_file, csv_file = NULL)

csv2midi(csv_file, midi_file = NULL)
}
\arguments{
\item{midi_file}{path of a midi file. Needs to be given for input in \code{midi2csv}.
If not given for output, a temporary file will be used for output}

\item{csv_file}{path of a csv file. Needs to be given for input in \code{csv2midi}.
If not given for output, a temporary file will be used}
}
\description{
Conversion between midi and csv files
\itemize{
\item \code{csv2midi()} converts from csv to midi
\item \code{midi2csv()} converts from midi to csv
}
}
\examples{
\dontrun{
midifile <- system.file("mididata/Monty_Python_s.mid",package="midicsvR")
csvfile <- "csv_from_midi.csv"
midi2csv(midifile,csvfile)
}
\dontrun{
csvfile <- system.file("mididata/Nokia_Tune.csv",package="midicsvR")
midifile <- "midi_from_csv.mid"
csv2midi(csvfile,midifile)
}

}
