% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/midi2csv.R
\name{midi2csv}
\alias{midi2csv}
\title{convert midi file to csv file}
\usage{
midi2csv(midi_file, csv_file)
}
\arguments{
\item{midi_file}{path of a midi file used as input}

\item{csv_file}{path of a csv file used as output}
}
\description{
\code{midi2csv} takes a midi file and outputs a textual representation
in a csv file
}
\examples{
\dontrun{
midifile <- system.file("mididata/Monty_Python_s.mid",package="midicsvR")
csvfile <- "csv_from_midi.csv"
midi2csv(midifile,csvfile)
}

}
