% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/midicsvR-package.R
\docType{package}
\name{midicsvR-package}
\alias{midicsvR}
\alias{midicsvR-package}
\title{midicsvR: Conversion between midi and csv files}
\description{
Midi files (a standard way to represent music as file) are binary files and cannot be handled easily in R. This package has functions to convert midi to csv, which can be read into R and transformed. The transformed csv files then can be saved as midi files. The package builds on C source code available at Fourmilab (https://www.fourmilab.ch/webtools/midicsv/) written by John Walker. This web site also has the documentation of the csv file format used.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/neuwirthe/midicsvR}
  \item Report bugs at \url{https://github.com/neuwirthe/midicsvR/issues}
}

}
\author{
\strong{Maintainer}: Erich Neuwirth \email{erich.neuwirth@univie.ac.at}

}
\keyword{internal}
